size_t usage for size related operations

+++ src/targetLK8000.cpp
@@ -180,7 +180,7 @@ void condor2nav::CTargetLK8000::TaskDump(CFileParserINI &profileParser,
   lk8000::WAYPOINT *startWaypointArray = new lk8000::WAYPOINT[lk8000::MAXSTARTPOINTS];
   memset(startWaypointArray, 0, lk8000::MAXSTARTPOINTS * sizeof(lk8000::WAYPOINT));
 
-  for(unsigned i=0; i<waypointArray.size(); i++) {
+  for(size_t i=0; i<waypointArray.size(); i++) {
     taskWaypointArray[i].Number = waypointArray[i].number;
     taskWaypointArray[i].Latitude = waypointArray[i].latitude;
     taskWaypointArray[i].Longitude = waypointArray[i].longitude;
@@ -270,7 +270,7 @@ void condor2nav::CTargetLK8000::Glider(const CFileParserCSV::CStringArray &glide
   polarFile << "*" << std::endl;
   polarFile << "* MassDryGross[kg], MaxWaterBallast[liters], Speed1[km/h], Sink1[m/s], Speed2, Sink2, Speed3, Sink3, WingArea[m2]" << std::endl;
   polarFile << "*****************************************************************************************************************" << std::endl;
-  for(unsigned i=GLIDER_MASS_DRY_GROSS; i<=GLIDER_SINK_3; i++) {
+  for(size_t i=GLIDER_MASS_DRY_GROSS; i<=GLIDER_SINK_3; i++) {
     if(i > GLIDER_MASS_DRY_GROSS)
       polarFile << ",";
     polarFile << gliderData.at(i);
